requirements.txt
meteostat>=1.6.0
pyyaml>=6.0
click>=8.1.0
numpy<2.0
